feat: Split functions into files for better imrprove order
